Update DynamicGraphTest to validate the order of free nodes.
The order of the nodes returned by getFreeNodes() is important - at least
other tests assert the order. However, no test in DynamicGraphTest tested
that ordering. This change makes all the tests in DynamicGraphTest assert
the order of the nodes returned by getFreeNodes() so the tests fail closer
to the implementation.